requirements.txt
alabaster==0.7.8
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==19.3.0
Automat==0.8.0
Babel==2.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
cachetools==5.2.0
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
cliapp==1.20180812.1
click==8.1.3
cloud-init==22.1
colorama==0.4.3
command-not-found==0.3
commonmark==0.9.1
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
docopt==0.6.2
docutils==0.16
entrypoints==0.3
executing==0.8.3
fastjsonschema==2.15.3
feedparser==5.2.1
Flask==2.1.2
gitdb==4.0.9
GitPython==3.1.27
gitsome==0.8.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
imagesize==1.2.0
importlib-metadata==4.12.0
incremental==16.10.1
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Markdown==3.1.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
more-itertools==4.2.0
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
netifaces==0.10.4
notebook==6.4.12
numpy==1.23.0
numpydoc==0.7.0
oauthlib==3.1.0
olefile==0.46
packaging==20.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==2.0.10
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==8.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.21
pydeck==0.7.1
Pygments==2.12.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
Pympler==1.0.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.3.1
pyzmq==23.2.0
requests==2.22.0
requests-unixsocket==0.2.0
rich==12.4.4
roman==2.0.0
SecretStorage==2.3.1
semver==2.13.0
Send2Trash==1.8.0
service-identity==18.1.0
setproctitle==1.1.10
simplejson==3.16.0
six==1.14.0
smmap==5.0.0
sos==4.3
soupsieve==2.3.2.post1
Sphinx==1.8.5
ssh-import-id==5.10
stack-data==0.3.0
streamlit==1.10.0
systemd-python==234
terminado==0.15.0
tinycss2==1.1.1
toml==0.10.2
toolz==0.11.2
tornado==6.1
traitlets==5.3.0
ttystatus==0.38
Twisted==18.9.0
typing-extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
ubuntu-advantage-tools==27.6
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
validators==0.20.0
wadllib==1.3.3
watchdog==2.1.9
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.1
xonsh==0.9.13
zipp==1.0.0
zope.interface==4.7.1
